+++ projects/safe/src/lib/components/ui/grid-core/grid-core.component.html
@@ -193,54 +193,6 @@
                 </ng-container>
             </kendo-grid-column-group>
         </ng-container> -->
-<!-- ROW DETAILS -->
-<!-- <kendo-grid-column *ngIf="showDetails" [columnMenu]="false" [width]="54" [resizable]="false">
-            <ng-template kendoGridCellTemplate let-dataItem="dataItem">
-                <button style="text-align: center" kendoButton [icon]="'detail-section'" [look]="'flat'"
-                    (click)="onShowDetails(dataItem)" matTooltip="Open detail"></button>
-            </ng-template>
-        </kendo-grid-column> -->
-<!-- ROW ACTIONS -->
-<!-- <kendo-grid-column *ngIf="!readOnly && hasEnabledActions" [columnMenu]="false" [width]="54" [resizable]="false">
-            <ng-template kendoGridCellTemplate let-dataItem="dataItem" let-idx="rowIndex">
-                <button
-                    *ngIf="settings.actions && ((settings.actions.update && dataItem.canUpdate) || settings.actions.history ||
-                    (settings.actions.convert && dataItem.canUpdate) || (settings.actions.delete && dataItem.canDelete))"
-                    style="text-align: center" kendoButton [icon]="'more-vertical'" [look]="'flat'"
-                    [matMenuTriggerFor]="menu">
-                </button>
-                <mat-menu #menu="matMenu">
-                    <button *ngIf="(!settings.actions || settings.actions.update) && dataItem.canUpdate"
-                        mat-menu-item color="warn" (click)="onUpdateRow(dataItem.id)">
-                        Update
-                    </button>
-                    <button *ngIf="!settings.actions || settings.actions.history" mat-menu-item color="warn"
-                        (click)="onViewHistory(dataItem.id)">
-                        History
-                    </button>
-                    <button *ngIf="(!settings.actions || settings.actions.convert) && dataItem.canUpdate"
-                        mat-menu-item color="warn" (click)="onConvertRecord([idx - skip])">
-                        Convert
-                    </button>
-                    <ng-container *ngIf="(!settings.actions || settings.actions.delete) && dataItem.canDelete">
-                        <mat-divider></mat-divider>
-                        <button mat-menu-item (click)="onDeleteRow([idx - skip])">
-                            Delete
-                        </button>
-                    </ng-container>
-                </mat-menu>
-            </ng-template>
-        </kendo-grid-column>
-    </ng-container> -->
-<!-- ROW SUBGRID -->
-<!-- <ng-container *ngIf="detailsField">
-        <div *kendoGridDetailTemplate="let dataItem">
-            <safe-grid-core [settings]="detailsField" [parent]="dataItem" (childChanged)="reloadData()"></safe-grid-core>
-        </div>
-    </ng-container> -->
-<!-- EXCEL EXPORT -->
-<!-- <kendo-grid-excel [fileName]="excelFileName"></kendo-grid-excel>
-</kendo-grid> -->
 <ng-template #toolbarTemplate>
     <div>test</div>
 </ng-template>
