+++ projects/safe/src/lib/components/widget-grid/floating-options/floating-options.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { SafeTileDisplayComponent } from './menu/tile-display/tile-display.component';
 import { SafeTileDataComponent } from './menu/tile-data/tile-data.component';
-import { SafeGridService } from '../../../services/grid.service';
 import { SafeConfirmModalComponent } from '../../confirm-modal/confirm-modal.component';
+import { SafeDashboardService } from '../../../services/dashboard.service';
 
 @Component({
   selector: 'safe-floating-options',
@@ -27,7 +27,7 @@ export class SafeFloatingOptionsComponent implements OnInit {
 
   constructor(
     public dialog: MatDialog,
-    private gridService: SafeGridService
+    private dashboardService: SafeDashboardService
   ) { }
 
   /*  Set the list of available actions.
@@ -71,7 +71,7 @@ export class SafeFloatingOptionsComponent implements OnInit {
       const dialogRef = this.dialog.open(SafeTileDataComponent, {
         data: {
           tile: this.widget,
-          template: this.gridService.findSettingsTemplate(this.widget)
+          template: this.dashboardService.findSettingsTemplate(this.widget)
         },
         // hasBackdrop: false,
         position: {
